requirements.txt
# ── Web Framework ──────────────────────────────────────────────
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2

# ── Kafka (Message Queue) ───────────────────────────────────────
kafka-python==2.0.2

# ── Task Queue & Cache ──────────────────────────────────────────
celery==5.4.0
redis==5.0.8

# ── Database ────────────────────────────────────────────────────
influxdb-client==1.44.0

# ── AI / ML ─────────────────────────────────────────────────────
scikit-learn==1.7.2
numpy==1.26.4
pandas==2.2.3

# ── Dashboard ───────────────────────────────────────────────────
streamlit==1.38.0

# ── Testing ─────────────────────────────────────────────────────
pytest==8.3.3
httpx==0.27.2
